Leftist heaps (wip).
Good progress so far, with many of the core functions already proved.
In this commit we intentionally use a preorder relation to compare
heap elements. Under such scenario, we are not able to complete the proof
of function `merge`. We take, however, this as a pedagogical opportunity
to elaborate on some details.

The implementation of `merge`, for the recursive case, is as follows:
```
  | N (_, x, a1, b1), N (_, y, a2, b2) ->
    if E.leq x y
    then _make_node x a1 (merge b1 t2)
    else _make_node y a2 (merge t1 b2)
```

If `x` is less or equal to `y`, we take the `then` branch we are
able to prove the precondition of `_make_node` that `le_root x a1`
and `le_root x (merge b1 t2)`. However, `x` is not less or equal to
`y`, we take `else` branch and are not able to prove the precondition
that `le_root y (merge t1 b2)` when calling `_make_node.

Why is that? The answer follows on the next commit.