Proof of leftist heap: using a total preorder.
When we take the `else` branch in the `merge` operation, we need to
know that in fact `y` is less or equal to `x`. If you take only a
preorder, we have no such knowledge. Moving to a total preorder only
amounts to adding the line
```
  (forall x y. leq x y /\ leq y x)
```
to the definition of the `pre_order` predicate. Using such a definition,
we are able to complete the proof of the `merge` operation.