Arith compiler: verified compilation of subtraction expressions.
We use a semantic function with errors for the evaluation of `OSub`
opcodes in our stack. Concretely, our semantics results are modeled as
follows:
```
  type result = Error | Res of stack
```
This resembles the use of `abort` results in Pretty-Big Step Semantics
by Arthur. This approach allows to us to provide a total definition
for the evaluation function of a list of opcodes. This makes the proof
of lemma `star_append` rather easy. However, we need to explicitly account
for the case of the `Error` result in intermediate evaluation steps.