Arith compiler: proved termination of all logical operations.
In order to prove the `star` and `star_append` definitions indeed
terminate, we introduce the auxiliary logical function `size_code`
which basically counts the number of nodes in the code component of
the state.

TODO: introduce a *fuel* measure, which must also be used to account
for divergent programs once we introduce `Loop` instructions.